cleanup: Move code in *-multipart-common.go to *-multipart.go
The code in *-multipart-common.go is not common anymore.
